backend/requirements.txt
fastapi==0.135.1
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.2.2
supabase>=2.0.0
uvicorn==0.41.0
